+++ aci_fabric_policies.tf
@@ -635,15 +635,21 @@ module "aci_user" {
 
 module "aci_login_domain" {
   source  = "netascode/login-domain/aci"
-  version = "0.2.0"
+  version = "0.2.1"
 
-  for_each    = { for dom in try(local.fabric_policies.aaa.login_domains, []) : dom.name => dom if local.modules.aci_login_domain && var.manage_fabric_policies }
-  name        = each.value.name
-  description = try(each.value.description, "")
-  realm       = try(each.value.realm, "")
+  for_each       = { for dom in try(local.fabric_policies.aaa.login_domains, []) : dom.name => dom if local.modules.aci_login_domain && var.manage_fabric_policies }
+  name           = each.value.name
+  description    = try(each.value.description, "")
+  realm          = try(each.value.realm, "")
+  auth_choice    = try(each.value.auth_choice, local.defaults.apic.fabric_policies.aaa.login_domains.auth_choice)
+  ldap_group_map = try(each.value.ldap_group_map, "")
   tacacs_providers = [for prov in try(each.value.tacacs_providers, []) : {
     hostname_ip = prov.hostname_ip
     priority    = try(prov.priority, local.defaults.apic.fabric_policies.aaa.login_domains.tacacs_providers.priority)
+  }]
+  ldap_providers = [for prov in try(each.value.ldap_providers, []) : {
+    hostname_ip = prov.hostname_ip
+    priority    = try(prov.priority, local.defaults.apic.fabric_policies.aaa.login_domains.ldap_providers.priority)
   }]
 
   depends_on = [
@@ -940,4 +946,45 @@ module "aci_fabric_span_source_group" {
   }]
   destination_name        = "${each.value.destination.name}${local.defaults.apic.fabric_policies.span.destination_groups.name_suffix}"
   destination_description = try(each.value.destination.description, "")
+}
+
+module "aci_ldap" {
+  source  = "netascode/ldap/aci"
+  version = "0.1.0"
+
+  ldap_providers = [for prov in try(local.fabric_policies.aaa.ldap.providers, []) : {
+    hostname_ip          = prov.hostname_ip
+    description          = try(prov.description, "")
+    port                 = try(prov.port, local.defaults.apic.fabric_policies.aaa.ldap.providers.port)
+    bind_dn              = try(prov.bind_dn, "")
+    base_dn              = try(prov.base_dn, "")
+    password             = try(prov.password, "")
+    timeout              = try(prov.timeout, local.defaults.apic.fabric_policies.aaa.ldap.providers.timeout)
+    retries              = try(prov.retries, local.defaults.apic.fabric_policies.aaa.ldap.providers.retries)
+    enable_ssl           = try(prov.enable_ssl, local.defaults.apic.fabric_policies.aaa.ldap.providers.enable_ssl)
+    filter               = try(prov.filter, "")
+    attribute            = try(prov.attribute, "")
+    ssl_validation_level = try(prov.ssl_validation_level, local.defaults.apic.fabric_policies.aaa.ldap.providers.ssl_validation_level)
+    mgmt_epg_type        = try(prov.mgmt_epg, local.defaults.apic.fabric_policies.aaa.ldap.providers.mgmt_epg)
+    mgmt_epg_name        = try(prov.mgmt_epg, local.defaults.apic.fabric_policies.aaa.ldap.providers.mgmt_epg) == "oob" ? try(local.node_policies.oob_endpoint_group, local.defaults.apic.node_policies.oob_endpoint_group) : try(local.node_policies.inb_endpoint_group, local.defaults.apic.node_policies.inb_endpoint_group)
+    monitoring           = try(prov.server_monitoring, local.defaults.apic.fabric_policies.aaa.ldap.providers.server_monitoring)
+    monitoring_username  = try(prov.monitoring_username, local.defaults.apic.fabric_policies.aaa.ldap.providers.monitoring_username)
+    monitoring_password  = try(prov.monitoring_password, "")
+  }]
+  group_map_rules = [for rule in try(local.fabric_policies.aaa.ldap.group_map_rules, []) : {
+    name        = rule.name
+    description = try(rule.description, "")
+    group_dn    = try(rule.group_dn, "")
+    security_domains = [for dom in try(rule.security_domains, []) : {
+      name = dom.name
+      roles = [for role in try(dom.roles, []) : {
+        name           = role.name
+        privilege_type = try(role.privilege_type, local.defaults.apic.fabric_policies.aaa.ldap.group_map_rules.security_domains.roles.privilege_type)
+      }]
+    }]
+  }]
+  group_maps = [for map in try(local.fabric_policies.aaa.ldap.group_maps, []) : {
+    name  = map.name
+    rules = [for rule in try(map.rules, []) : rule.name]
+  }]
 }